+++ sqldelight-idea-plugin/src/main/kotlin/app/cash/sqldelight/intellij/resolvers/SqlDelightProjectResolverExtension.kt
@@ -1,5 +1,6 @@
 package app.cash.sqldelight.intellij.resolvers
 
+import app.cash.sqldelight.core.GradleCompatibility
 import app.cash.sqldelight.core.SqlDelightCompilationUnit
 import app.cash.sqldelight.core.SqlDelightDatabaseName
 import app.cash.sqldelight.core.SqlDelightDatabaseProperties
@@ -39,6 +40,7 @@ class SqlDelightProjectResolverExtension : AbstractProjectResolverExtension() {
             databaseProperties.treatNullAsUnknownForEquality,
             databaseProperties.generateAsync,
             databaseProperties.rootDirectory,
+            GradleCompatibility.readExpandSelectStar(sqlDelightModel, databaseProperties),
           )
         },
         sqlDelightModel.dialectJars.toMutableList(),
@@ -73,6 +75,7 @@ data class SqlDelightDatabasePropertiesModel(
   override val treatNullAsUnknownForEquality: Boolean = false,
   override val generateAsync: Boolean = false,
   override val rootDirectory: File,
+  override val expandSelectStar: Boolean = true,
 ) : SqlDelightDatabaseProperties
 
 data class SqlDelightDatabaseNameModel(