+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## Unreleased
 
+Nothing here yet!
+
+## [2.3.2] - 2026-03-16
+[2.3.2]: https://github.com/sqldelight/sqldelight/releases/tag/2.3.2
+
 ### Added
 - [PostgreSQL Dialect] Improved support for ALTER TABLE ALTER TYPE USING expressions (#6116 by @griffio)
 - [PostgreSQL Dialect] Add support for DROP COLUMN IF EXISTS (#6112 by @griffio)
+- [Gradle Plugin] Add expandSelectStar flag to switch off Select wildcard expansions (#5813 by @griffio)
 - [MySQL Dialect] Add support for Window Functions (#6086 by @griffio)
 - [Gradle Plugin] Fix build failure when starting schema version is not 1 and verifyMigrations is true (#6017 by @neilgmiller)
 - [Gradle Plugin] Make `SqlDelightWorkerTask` more configurable, and update default configuration to support developing on Windows (#5215 by @MSDarwish2000)
@@ -16,12 +22,14 @@
 - [Gradle Plugin] Full compatibility with Android Gradle Plugin 9.0's new DSL. (#6140)
 - [PostgreSQL Dialect] Support PostgreSql CREATE TABLE storage parameters (#6148 by @griffio)
 - [PostgreSQL Dialect] Add query support for implicitly defined System Columns (#5834 by @griffio)
+- [PostgreSQL Dialect] Fix PostgreSql unique table constraint nullable result column (#6167 by @griffio)
 
 ### Changed
 - [Compiler] Change compiler output type from java.lang.Void to kotlin.Nothing (#6099 by @griffio)
 - [Compiler] Allow for underscores in package names. Before underscores have been sanitized which resulted in unexpected behaviours (#6027 by @BierDav)
 - [Paging Extension] Switch to AndroidX Paging (#5910 by @jeffdgr8)
 - [Android Driver] Increase Android minSdk to 23. (#6141)
+- [Paging Extension] Upgrade to paging 3.4.1, and removed X64 apple targets. (#6166)
 
 ### Fixed
 - [IntelliJ Plugin] Fix IDE freeze caused by blocking file type detection on the EDT during VFS refresh events.
@@ -30,6 +38,17 @@
 - [Gradle Plugin] Improve performance of `VerifyMigrationTask` to stop it from hanging on complex schemas (#6073 by @Lightwood13)
 - [Intellij Plugin] Fix Plugin initialization exceptions and update deprecated methods (#6040 by @griffio)
 - [Gradle Plugin] Fix compatibility with Android Gradle Plugin's built-in Kotlin (#6139)
+
+## [2.3.1] - 2025-03-12
+[2.3.1]: https://github.com/sqldelight/sqldelight/releases/tag/2.3.1
+
+Failed release. Use 2.3.2!
+
+
+## [2.3.0] - 2025-03-12
+[2.3.0]: https://github.com/sqldelight/sqldelight/releases/tag/2.3.0
+
+Failed release. Use 2.3.2!
 
 ## [2.2.1] - 2025-11-13
 [2.2.1]: https://github.com/sqldelight/sqldelight/releases/tag/2.2.1
@@ -50,6 +69,7 @@
 - [Gradle Plugin] Add support for version catalogs when adding modules (#5755 by @DRSchlaubi)
 
 ### Changed
+- [PostgreSQL Dialect] Change arrayIntermediateType visibility to public (#5835 by [Griffio][griffio])
 - In-development snapshots are now published to the Central Portal Snapshots repository at https://central.sonatype.com/repository/maven-snapshots/.
 - [Compiler] Simplified default generated queries using constructor references (#5814 by @jonapoul)